Refactored User classes into Model folder and activities into controller folder.
