+++ app/src/main/java/edu/gatech/cs2340/hkskh/Controller/LoginActivity.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs2340.hkskh;
+package edu.gatech.cs2340.hkskh.Controller;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -7,6 +7,8 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import edu.gatech.cs2340.hkskh.R;
 
 public class LoginActivity extends AppCompatActivity {
