+++ app/src/main/java/edu/gatech/cs2340/hkskh/Controller/MainActivity.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs2340.hkskh;
+package edu.gatech.cs2340.hkskh.Controller;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -8,7 +8,7 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import org.w3c.dom.Text;
+import edu.gatech.cs2340.hkskh.R;
 
 public class MainActivity extends AppCompatActivity {
