+++ app/src/main/java/edu/gatech/cs2340/hkskh/Controller/WelcomeActivity.java
@@ -1,10 +1,12 @@
-package edu.gatech.cs2340.hkskh;
+package edu.gatech.cs2340.hkskh.Controller;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import edu.gatech.cs2340.hkskh.R;
 
 public class WelcomeActivity extends AppCompatActivity {
